REFACTOR(RestApi proj): moved bootstrapping code to "OnlineMinion.RestApi.Init" namespace
