+++ OnlineMinion.RestApi/Init/ApiBehaviorOptionsConfigurator.cs
@@ -1,7 +1,7 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Options;
 
-namespace OnlineMinion.RestApi.Configuration;
+namespace OnlineMinion.RestApi.Init;
 
 public class ApiBehaviorOptionsConfigurator : IConfigureOptions<ApiBehaviorOptions>
 {
