+++ OnlineMinion.RestApi/Init/ApiExplorerOptionsConfigurator.cs
@@ -1,7 +1,7 @@
 using Asp.Versioning.ApiExplorer;
 using Microsoft.Extensions.Options;
 
-namespace OnlineMinion.RestApi.Configuration;
+namespace OnlineMinion.RestApi.Init;
 
 public class ApiExplorerOptionsConfigurator : IConfigureOptions<ApiExplorerOptions>
 {
